BROKEN-ish: Implement postprocess()
I think I got postprocess() working! The code is really ugly and needs clean-up. It technically gives a different output than it used to. Previooulsy the entity "Conservative People ' s Party of Estonia" had a confidence of 0.983; now it has a confidence of only 0.876. I think this is because I started de-meaning the logits (which is what the parent class does).